% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCA.R
\name{cnonct}
\alias{cnonct}
\title{Non-centrality parameter for chi-square distribution}
\usage{
cnonct(df, alpha, beta)
}
\arguments{
\item{df}{an integer giving the degrees of freedom of the chi-square variable}

\item{alpha}{a numeric value giving the significance level of the test}

\item{beta}{a numeric value giving the desired type II error (1-\code{beta} is the power)}
}
\description{
Calculates the non-centrality parameter for a chi-square distribution that achieves the target power at a given significance level. This is often needed for sample size calculation for chi-square based tests.
}
\examples{
cnonct(6, 0.05, 0.2)
}

