% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCA.R
\name{internal}
\alias{.bit2boolean}
\alias{.p.adjust.closed}
\alias{internal}
\title{Internal functions}
\usage{
.bit2boolean(x, N)

.p.adjust.closed(test, hypotheses, remove = FALSE, ...)
}
\arguments{
\item{test}{function that performs the local test. The function should accept a subvector of the hypotheses argument as input, and return a p-value.}

\item{hypotheses}{identifiers of the collection of elementary hypotheses.}

\item{remove}{logical indicator of whether hypotheses of length N-1 should be removed}

\item{...}{additional parameters to the 'test' function}
}
\value{
numeric vector of adjusted p-values for each hypothesis
}
\description{
These internal functions perform the closed set p-value adjustment calculation
for the multivariate Cochran-Armitage trend test. The logical constraint
on the possible number of true null hypotheses is incorporated.
}
\keyword{internal}

