\name{MultiCA-methods}
\alias{print.mbca}
\alias{print.bcatest}
\alias{plot.mbca}

\title{MultiCA print and plot methods}

\description{Printing and plotting objects in the MultiCA package.}

\usage{
\method{print}{bcatest}(x, ...)
}

\arguments{
  \item{x}{An object of class mbca or bcatest.}
  \item{...}{Further arguments.}
}


\author{Daniel Gerhard}

\keyword{ methods }
