\name{multiCA-package}
\alias{multiCA-package}
\alias{multiCA}
\docType{package}
\title{
\packageTitle{multiCA}
}
\description{
\packageDescription{multiCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multiCA}
\packageIndices{multiCA}

The main functionality is implemented in the \code{multiCA.test} function. The \code{power.multiCA.test} function can be used for power and sample size calculation.
}
\author{
\packageAuthor{multiCA}

Maintainer: \packageMaintainer{multiCA}
}
\references{
Szabo, A. (2016) Test for trend with a multinomial outcome.
}

\keyword{ package
htest
nonparametric}

