% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCA.R
\name{power.CA.test}
\alias{power.CA.test}
\title{Power calculations for the Cochran-Armitage trend test}
\usage{
power.CA.test(
  N = NULL,
  power = NULL,
  pvec = NULL,
  scores = seq_along(pvec),
  n.prop = rep(1, length(pvec)),
  sig.level = 0.05
)
}
\description{
Power calculations for the Cochran-Armitage trend test
}
